Trait emtm_db::controller::user_controller::UserController

pub trait UserController {
    fn add_cows(&self, cows: Vec<Cow>) -> Vec<Result<i32, String>>;

    fn add_students(&self, students: Vec<Student>) -> Vec<Result<i32, String>>;

    fn add_db_user(&self, new_user: NewUser) -> Result<User, String>;

    fn get_user_from_uid(&self, to_search: i32) -> Option<User>;

    fn get_user_from_username(&self, name: &str) -> Option<User>;

    fn get_cow_from_uid(&self, to_search: i32) -> Option<Cow>;

    fn get_student_from_uid(&self, to_search: i32) -> Option<Student>;
}

Required methods

fn add_cows(&self, cows: Vec<Cow>) -> Vec<Result<i32, String>>

Add cows with a vector of Cows, returns the uid of added cows

Arguments

'cows' - A vector containing all the cow instances to be added

fn add_students(&self, students: Vec<Student>) -> Vec<Result<i32, String>>

Add students with a vector of Students, returns the uid of added students

Arguments

'students' - A vector containing all the student instances to be added

fn add_db_user(&self, new_user: NewUser) -> Result<User, String>

Helper function for add_cows and add_db_user, returns the added user

Arguments

'new_user' - A vector containing the user to be added

fn get_user_from_uid(&self, to_search: i32) -> Option<User>

Query a user from a given uid

Arguments

'to_search' - The uid to be search

fn get_user_from_username(&self, name: &str) -> Option<User>

Query a user from a given username

Arguments

'name' - The username to be search

fn get_cow_from_uid(&self, to_search: i32) -> Option<Cow>

Query a cow from a given uid

Arguments

'to_search' - The uid to be search

fn get_student_from_uid(&self, to_search: i32) -> Option<Student>

Query a cow from a given uid

Arguments

'to_search' - The uid to be search

Loading content...